@@ -100,7 +100,7 @@
     "debounce-promise": "^3.1.2",
     "debug-menu": "^0.6.1",
     "document-register-element": "^1.14.3",
-    "electron": "^6.1.5",
+    "electron": "^7.1.2",
     "electron-devtools-installer": "^2.2.4",
     "fs-extra": "^8.1.0",
     "hyperhtml": "^2.31.5",
