@@ -2,153 +2,17 @@ import * as webpack from 'webpack';
 import * as path from 'path';
 
 import { CleanWebpackPlugin } from 'clean-webpack-plugin';
-import * as HtmlWebpackPlugin from 'html-webpack-plugin';
-// import * as WebpackBuildNotifierPlugin from 'webpack-build-notifier';
 import * as WebpackNotifierPlugin from 'webpack-notifier';
 import webpackMerge = require('webpack-merge');
 
 import configBase, { PROJECT_PATH } from './webpack.config.base';
 import { BuildEnvironment } from './types';
 
-function createRendererConfig(env: BuildEnvironment): webpack.Configuration {
-
-  // const include = [
-  //   fs.realpathSync(path.join(PROJECT_PATH, './src/shared/maestro/')),
-  // ];
-
-  // ?path=http://localhost:${PORT}/__webpack_hmr&reload=true&timeout=20000
-  // http://localhost:${PORT}/
-  // const PORT = process.env.PORT || 3000;
-  const HOT_MW = `webpack-hot-middleware/client?path=/__webpack_hmr&reload=true&timeout=20000`;
-  const ENTRY_PATH = path.join(PROJECT_PATH, './src/renderer');
-
-  const LAUNCHER_PATH = path.join(ENTRY_PATH, '/launcher');
-  const EDITOR_PATH = path.join(ENTRY_PATH, '/editor');
-
-  const LAUNCHER_ENTRY_PATH = path.join(LAUNCHER_PATH, '/index.ts');
-  const EDITOR_ENTRY_PATH = path.join(EDITOR_PATH, '/index.ts');
-
-
-  const config = webpackMerge.smart(
-    configBase('renderer', ENTRY_PATH, env),
-    {
-      target: 'electron-renderer',
-      entry: {
-        'launcher': (env.isProduction) ? LAUNCHER_ENTRY_PATH : [HOT_MW, LAUNCHER_ENTRY_PATH],
-        'editor': (env.isProduction) ? EDITOR_ENTRY_PATH : [HOT_MW, EDITOR_ENTRY_PATH],
-      },
-      output: {
-        publicPath: `/dist/renderer`,
-        filename: '[name]/index.bundle.js',
-        hotUpdateChunkFilename: ".hot/[id].[hash].hot-update.js",
-        hotUpdateMainFilename: ".hot/[hash].hot-update.json"
-      },
-      module: {
-        rules: [
-          // {
-          //   test: /\.ts$/,
-          //   enforce: "pre",
-          //   loader: 'eslint-loader',
-          //   exclude: /node_modules/,
-          // },
-          {
-            test: /\.ts$/,
-            loader: 'ts-loader',
-            exclude: /node_modules/
-          },
-          {
-            test: /\.p?css$/,
-            exclude: /node_modules/,
-            use: [
-              {
-                loader: 'style-loader',
-              },
-              {
-                loader: 'css-loader',
-                options: {
-                  importLoaders: 1,
-                  url: true,
-                }
-              },
-              {
-                loader: 'postcss-loader',
-                options: {
-                  ident: 'postcss',
-                  config: {
-                    path: __dirname,
-                  }
-                }
-              },
-            ]
-          },
-          {
-            test: /\.(jpe?g|png|gif)$/i,
-            loaders: [
-              {
-                loader: 'file-loader',
-                options: {
-                  useRelativePaths: true,
-                  name: (env.isProduction) ? '[sha512:hash:hex:9].[ext]' : '[name].[ext]',
-                  emitFile: true,
-                  outputPath: (url: string, resourcePath: string): string => {
-                    const chunk = /launcher|editor/g.exec(resourcePath);
-                    return `${chunk[0]}/img/${url}`;
-                  }
-                }
-              },
-              // 'file-loader?hash=sha512&digest=hex&name=img/[hash].[ext]',
-              // 'image-webpack-loader?bypassOnDebug&optipng.optimizationLevel=7&gifsicle.interlaced=false',
-            ],
-          },
-          {
-            test: /.svg$/i,
-            loader: 'svg-inline-loader'
-          }
-        ],
-      },
-      plugins: [
-        new webpack.DefinePlugin({
-          'DEVELOPMENT': JSON.stringify(
-            true
-          ),
-        }),
-        new WebpackNotifierPlugin({
-          title: "Purpurina <Renderer>",
-          alwaysNotify: true,
-          // logo: path.resolve("./img/favicon.png"),
-        }),
-        new CleanWebpackPlugin(),
-        new HtmlWebpackPlugin({
-          filename: "launcher/index.html",
-          inject: 'body',
-          chunks: ['launcher'],
-          template: path.join(LAUNCHER_PATH, `./index.html`),
-          minify: false,
-        }),
-        new HtmlWebpackPlugin({
-          filename: "editor/index.html",
-          inject: 'body',
-          chunks: ['editor'],
-          template: path.join(EDITOR_PATH, `./index.html`),
-          minify: false,
-        }),
-      ]
-    });
-
-  if (!env.isProduction) {
-    config.plugins.push(
-      new webpack.HotModuleReplacementPlugin(),
-      new webpack.NoEmitOnErrorsPlugin()
-    );
-  }
-  return config;
-}
+import rendererConfig from './webpack.config.renderer';
 
 export default (env: BuildEnvironment): { main: webpack.Configuration; renderer: webpack.Configuration } => {
 
   const MAIN_ENTRY_PATH = path.join(PROJECT_PATH, './src/main');
-  // const RENDER_ENTRY_PATH = path.join(PROJECT_PATH, './src/renderer');
-
   return {
     main: webpackMerge.smart(
       configBase('', MAIN_ENTRY_PATH, env),
@@ -158,7 +22,10 @@ export default (env: BuildEnvironment): { main: webpack.Configuration; renderer:
           main: path.join(MAIN_ENTRY_PATH + '/index.ts')
         },
         resolve: {
-          mainFields: ["electron-main", "module", "main"]
+          mainFields: ["electron-main", "module", "main"],
+          alias: {
+            '@main': path.join(PROJECT_PATH, `./src/main`),
+          }
         },
         plugins: [
           new WebpackNotifierPlugin({
@@ -169,7 +36,7 @@ export default (env: BuildEnvironment): { main: webpack.Configuration; renderer:
           new webpack.NoEmitOnErrorsPlugin(),
         ]
       }),
-    renderer: createRendererConfig(env),
+    renderer: rendererConfig(env),
   }
 }
 
