@@ -6,7 +6,6 @@ import stars = require('../../img/icon_stars.svg');
 import learn = require('../../img/icon_learn.svg');
 import home = require('../../img/icon_home.svg');
 import settings = require('../../img/icon_settings.svg');
-
 import iconDefs = require('../../img/icon_defs.svg');
 
 interface MenuState {
@@ -26,7 +25,7 @@ export default class Navigation extends hyper.Component<MenuState> {
   }
 
   private onMenuSelect = (id: number, e: MouseEvent): void => {
-    const target = e.target as HTMLElement;
+    // const target = e.target as HTMLElement;
     // e.preventDefault();
     (document.activeElement as HTMLElement).blur();
     // target.focus({ preventScroll: true });
